Added placeholder functions to all distributions in x.random.hpp for functions that characterize the distributions.
Each class now has functions for:
- pdf
- cdf
- mean
- median
- mode
- variance
- skewness
- kurtosis

There are flags for the moments (meandef, mediandef, etc)
to check to see if they're defined or not.

Since only one of the distribution classes has actual values for these
(uniform_int_distribution), there is a temporary flag in the Max object
which notes whether these functions have valid values or not. If so,
then messages are added for them when the Max object is instantiated,
and if not, they are left out.