Added a validparams message to all classes and support for that in the Max objects.
validparams returns a struct containing a true / false field and a
string with a message describing why the value is false, or the
empty string if it's true.

The rationale behind this implementation over a more heavy-handed
one that reports invalid parameter values as they are encountered
is that for many distributions, because parameters depend on each
other and must be set sequentially, there are often times where
parameter sets are invalid, but are in the process of changing,
so we don't want overly aggressive error reporting.

Second, although some parameter values are technically invalid,
they may still produce useful results in some contexts
(e.g. a multinomial with 0 trials).

So, the validparams function, and corresponding Max message
are there to be used if / when appropriate, and ignored if desired.